// frontend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        RecipeRead: {
            /** Title */
            title: string;
            /** Description */
            description?: string | null;
            /** Prep Time */
            prep_time?: number | null;
            /** Cook Time */
            cook_time?: number | null;
            /** Servings */
            servings?: number | null;
            /** Image Url */
            image_url?: string | null;
            /** Source */
            source?: string | null;
            /** Source Url */
            source_url?: string | null;
            /** Categories */
            categories?: string[] | null;
            /** Tags */
            tags?: string[] | null;
            /** Difficulty */
            difficulty?: string | null;
            /** Notes */
            notes?: string | null;
            /** Rating */
            rating?: number | null;
            /** Nutritional Info */
            nutritional_info?: string | null;
            /**
             * Ingredients
             * @default []
             */
            ingredients: components["schemas"]["RecipeIngredientRead"][];
            /** Instructions */
            instructions: components["schemas"]["RecipeInstructionRead"][];
            nutrition?: components["schemas"]["RecipeNutritionRead"] | null;
            /** Id */
            id: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}